Update lightCurve.py
Updated the lightCurve plotting function to be more general to work with a more general data set; in particular, the previous version assumes the oid changes at particular locations in the list, and the software of this new version determines that. This version still assumes there are at least two unique object id's and no more than three.