Update lightCurve.py
There are sources with only one object id and others with >3, and Adam said he will merge with the repo when a more general version of the software is available. This version of the lightCurve software is the fully general version; it works for sources with any number of object ID's!